move init_file into subpackage
